\name{cols}
\alias{cols}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Function that returns the color scale for 'image()'.
}
\description{
Function that returns the color scale for 'image()'. This function is a
simplification of the function 'tim.colors()' from the 'fields' package.
}
\usage{
cols()
}

\value{
A vector with colors.
}
\references{
Fields Development Team (2006). fields: Tools for Spatial Data. National Center for Atmospheric Research, Boulder, CO. URL http://www.cgd.ucar.edu/Software/Fields. 
}
\author{
Fabio Sigrist
}
