\name{index.complex.to.real.dft}
\alias{index.complex.to.real.dft}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
 Auxilary function for the real Fourier transform.
}
\description{
Auxilary function for the conversion between the complex FFT and the real Fourier transform.
}
\usage{
index.complex.to.real.dft(n)
}
\arguments{
  \item{n}{
 Number of points on each axis. n x n is the total number of spatial points.
}

}
\value{
  A a list of indices used for the conversion between the complex FFT and the real Fourier transform.

}
\author{
Fabio Sigrist
}
