\name{print.spateMCMC}
\alias{print.spateMCMC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print function for spateMCMC objects.
}
\description{
Print function for spateMCMC objects.
}
\usage{
\method{print}{spateMCMC}(x,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
    A 'spateMCMC' object obtained from 'spate.mcmc'.
  }
  \item{...}{not used.}
}

\author{
Fabio Sigrist
}

