\name{vnorm}
\alias{vnorm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Eucledian norm of a vector
}
\description{
Calculates the Eucledian norm of a vector
}
\usage{
vnorm(v)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{v}{
Vector.
}
}

\value{
The Eucledian norm of the vector 'v'.
}
\author{
Fabio Sigrist
}

\examples{
v <- c(1,2)
vnorm(v)
}
